+++ packages/std/src/math/mod.rs
@@ -1,5 +1,7 @@
 mod decimal;
+mod fraction;
 mod uint128;
 
 pub use decimal::Decimal;
+pub use fraction::Fraction;
 pub use uint128::Uint128;